* use `nuxtApp.provide` instead of `useState` for singleton pattern for Hanko instance
