+++ playground/pages/about.vue
@@ -2,7 +2,7 @@
   <main>
     <h1>About</h1>
     <p>
-      This page is accesible to both logged in and logged out users. No middleware is defined for
+      This page is accessible to both logged in and logged out users. No middleware is defined for
       this page.
     </p>
   </main>
