+++ playground/pages/index.vue
@@ -4,7 +4,7 @@
   <main>
     <h1>Home</h1>
     <p>
-      This page is accesible to both logged in and logged out users. No middleware is defined for
+      This page is accessible to both logged in and logged out users. No middleware is defined for
       this page. Use the navigation above to navigate to other pages.
     </p>
   </main>
