+++ src/runtime/plugins/components.client.ts
@@ -1,5 +1,17 @@
-import { register } from '@teamhanko/hanko-elements'
+import { Hanko, register } from '@teamhanko/hanko-elements'
 import { defineNuxtPlugin, useRuntimeConfig } from '#imports'
+
+declare module '#app' {
+  interface NuxtApp {
+    $hanko: InstanceType<typeof Hanko> | null
+  }
+}
+
+declare module 'vue' {
+  interface ComponentCustomProperties {
+    $hanko: InstanceType<typeof Hanko> | null
+  }
+}
 
 export default defineNuxtPlugin((nuxtApp) => {
   const config = useRuntimeConfig()
@@ -13,4 +25,12 @@ export default defineNuxtPlugin((nuxtApp) => {
       ...config.public.hanko.components,
     })
   })
+
+  const hankoConfig = useRuntimeConfig().public.hanko
+  nuxtApp.provide('hanko', new Hanko(hankoConfig.apiURL, {
+    cookieName: hankoConfig.cookieName,
+    localStorageKey: hankoConfig.storageKey,
+    cookieSameSite: hankoConfig.cookieSameSite,
+    cookieDomain: hankoConfig.cookieDomain,
+  }))
 })
